[14.0][IMP] l10n_it_fatturapa_out: make template data overridable
