+++ l10n_it_fatturapa_out/wizard/efattura.py
@@ -52,6 +52,18 @@ class FPAValidator:
 DEFAULT_INVOICE_ITALIAN_DATE_FORMAT = "%Y-%m-%d"
 
 
+def format_numbers(number):
+    # format number to str with between 2 and 8 decimals (event if it's .00)
+    number_splited = str(number).split(".")
+    if len(number_splited) == 1:
+        return "%.02f" % number
+
+    cents = number_splited[1]
+    if len(cents) > 8:
+        return "%.08f" % number
+    return float_repr(number, max(2, len(cents)))
+
+
 class EFatturaOut:
 
     _validator = FPAValidator()
@@ -75,17 +87,6 @@ class EFatturaOut:
             # Format the monetary values to avoid trailing decimals
             # (e.g. 90.85000000000001).
             return float_repr(number, min(2, currency.decimal_places))
-
-        def format_numbers(number):
-            # format number to str with between 2 and 8 decimals (event if it's .00)
-            number_splited = str(number).split(".")
-            if len(number_splited) == 1:
-                return "%.02f" % number
-
-            cents = number_splited[1]
-            if len(cents) > 8:
-                return "%.08f" % number
-            return float_repr(number, max(2, len(cents)))
 
         def format_numbers_two(number):
             # format number to str with 2 (event if it's .00)
@@ -195,67 +196,9 @@ class EFatturaOut:
             return False
 
         def get_all_taxes(record):
-            """Generate summary data for taxes.
-            Odoo does that for us, but only for nonzero taxes.
-            SdI expects a summary for every tax mentioned in the invoice,
-            even those with price_total == 0.
-            """
-
-            def _key(tax_id):
-                return tax_id.id
-
-            out_computed = {}
-            # existing tax lines
-            tax_ids = record.line_ids.filtered(lambda line: line.tax_line_id)
-            for tax_id in tax_ids:
-                tax_line_id = tax_id.tax_line_id
-                aliquota = format_numbers(tax_line_id.amount)
-                key = _key(tax_line_id)
-                out_computed[key] = {
-                    "AliquotaIVA": aliquota,
-                    "Natura": tax_line_id.kind_id.code,
-                    # 'Arrotondamento':'',
-                    "ImponibileImporto": tax_id.tax_base_amount,
-                    "Imposta": tax_id.price_total,
-                    "EsigibilitaIVA": tax_line_id.payability,
-                }
-                if tax_line_id.law_reference:
-                    out_computed[key]["RiferimentoNormativo"] = encode_for_export(
-                        tax_line_id.law_reference, 100
-                    )
-
-            out = {}
-            # check for missing tax lines
-            for line in record.invoice_line_ids:
-                if line.display_type in ("line_section", "line_note"):
-                    # notes and sections
-                    # we ignore line.tax_ids altogether,
-                    # (it is popolated with a default tax usually)
-                    # and use another tax in the template
-                    continue
-                for tax_id in line.tax_ids:
-                    aliquota = format_numbers(tax_id.amount)
-                    key = _key(tax_id)
-                    if key in out_computed:
-                        continue
-                    if key not in out:
-                        out[key] = {
-                            "AliquotaIVA": aliquota,
-                            "Natura": tax_id.kind_id.code,
-                            # 'Arrotondamento':'',
-                            "ImponibileImporto": line.price_subtotal,
-                            "Imposta": 0.0,
-                            "EsigibilitaIVA": tax_id.payability,
-                        }
-                        if tax_id.law_reference:
-                            out[key]["RiferimentoNormativo"] = encode_for_export(
-                                tax_id.law_reference, 100
-                            )
-                    else:
-                        out[key]["ImponibileImporto"] += line.price_subtotal
-                        out[key]["Imposta"] += 0.0
-            out.update(out_computed)
-            return out
+            # wrapper to a method in wizard (for better overriding)
+            wiz = self.env["wizard.export.fatturapa"]
+            return wiz.getAllTaxes(record)
 
         def get_importo(line):
             str_number = str(line.discount)
@@ -268,6 +211,11 @@ class EFatturaOut:
             # wrapper to a method in wizard (for better overriding)
             wiz = self.env["wizard.export.fatturapa"]
             return wiz.getImportoTotale(invoice)
+
+        def get_payments(invoice):
+            # wrapper to a method in wizard (for better overriding)
+            wiz = self.env["wizard.export.fatturapa"]
+            return wiz.getPayments(invoice)
 
         def fpa_to_eur(amount, invoice):
             currency = invoice.currency_id
@@ -310,12 +258,15 @@ class EFatturaOut:
             "wizard": self.wizard,
             "get_importo": get_importo,
             "get_importo_totale": get_importo_totale,
+            "get_payments": get_payments,
             "all_taxes": {
                 invoice.id: get_all_taxes(invoice) for invoice in self.invoices
             },
             "fpa_to_eur": fpa_to_eur,
         }
-        return template_values
+
+        wiz = self.env["wizard.export.fatturapa"]
+        return wiz.getTemplateValues(template_values)
 
     def to_xml(self, env):
         """Create the xml file content.
